Refactored configurations module
