+++ components/bpmn-q/modeler-component/editor/configurations/ConfigurationEndpoint.js
