+++ components/bpmn-q/modeler-component/editor/configurations/Constants.js
