--- components/bpmn-q/modeler-component/extensions/configurations-extension/ConfigurationsPlugin.js
@@ -1,6 +1,0 @@
-let dataflowModdleDescriptor = require('./resources/configurations-extension.json');
-
-export default {
-  name: 'configs',
-  moddleDescription: dataflowModdleDescriptor,
-};
