--- components/bpmn-q/modeler-component/extensions/configurations-extension/resources/configurations-extension.json
@@ -1,22 +1,0 @@
-{
-  "name": "ConfigurationsExtension",
-  "uri": "https://github.com/configurations/",
-  "prefix": "configs",
-  "xml": {
-    "tagAlias": "lowerCase"
-  },
-  "types": [
-    {
-      "name": "ConfigurableElement",
-      "properties": [
-        {
-          "name": "selectedConfigurationId",
-          "isAttr": true,
-          "type": "string"
-        }
-      ]
-    }
-  ],
-  "enumerations": [],
-  "associations": []
-}
