+++ components/bpmn-q/modeler-component/extensions/data-extension/configurations/TransformationTaskConfigurations.js
@@ -1,4 +1,4 @@
-import ConfigurationsEndpoint from '../../configurations-extension/configurations/ConfigurationEndpoint';
+import ConfigurationsEndpoint from '../../../editor/configurations/ConfigurationEndpoint';
 import * as consts from '../Constants';
 
 const endpoint = new ConfigurationsEndpoint('http://localhost:8000/service-task');
