+++ components/bpmn-q/modeler-component/extensions/data-extension/resources/data-flow-extension.json
@@ -30,7 +30,7 @@
     },
     {
       "name": "TransformationTask",
-      "superClass": [ "bpmn:Task", "configs:ConfigurableElement" ],
+      "superClass": [ "bpmn:Task" ],
       "properties": [
         {
           "name": "parameters",
