+++ components/bpmn-q/modeler-component/extensions/quantme/configurations/DataObjectConfigurations.js
@@ -1,4 +1,4 @@
-import ConfigurationsEndpoint from '../../configurations-extension/configurations/ConfigurationEndpoint';
+import ConfigurationsEndpoint from '../../../editor/configurations/ConfigurationEndpoint';
 import * as dataConsts from '../../data-extension/Constants';
 import * as configManager from '../framework-config/config-manager';
